Kernel: Rename File(Description)::can_write > can_write_without_blocking
According to Andreas, this function is supposed to answer "can I write
to this file (description) without blocking?"

This renames the function to reflect that, as it's not obvious from
just looking at the name and may be interpreted as "can I write to this
at all?"

This seems to have happened in some places because they return false
for this function and then always return an error on an attempt to
write.

This commit does not fix these areas and will be done in a separate
commit.