Kernel: Make read/write non-blocking if not supported/implemented
A bunch of files were making these operations blocking because they
were returning false in can_{read,write}_without_blocking. This is
because they do not support one or both of these operations and the
original names of the functions were misleading.

This also makes the error code returned in read/write consistent.
It returns ENOTSUP if not supported at all and ENOTIMPL if not
implemented just yet.

Also updates the comment in File.h to reflect these changes.