LibWeb: Add CallWithGlobalObject custom IDL attribute
Adding this attribute to a function causes the global object to be
passed as the first argument to the C++ function.

This is especially needed by static IDL functions which have little to
no access to the state around them (as they're implemented as C++
static functions). For example, the AbortSignal.abort static function
requires the global object to wrap an AbortError DOMException.