LibJS: Map bytecode instructions to their source code range
This is done by creating a "debug block" per AST node. This debug block
contains the start and end pointers into the instruction stream for the
instructions associated with that AST node and a copy of it's shrink
wrapped source range. It is done as a copy to allow the AST to be torn
down in the future, as it keeps no references to the AST.

To create the debug blocks and populate them with correct information,
a new RAII helper is introduced called "GeneratorNodeScope" which
operates in a very similar fashion to InterpreterNodeScope. The helper
should be placed at the beginning of the generation of each AST node.

To map an instruction to source code, you must find it's "closest"
debug block. This is done by finding the smallest, most recent debug
block that has the instruction's pointer in range. As an example, this
is done by linear search per instruction when dumping an executable,
which can definitely be done more efficiently. :^)